+++ lib/infrastructure/http/circle_api/circle_api_service.dart
@@ -1,5 +1,5 @@
 import 'package:dio/dio.dart';
-import 'package:m3_app/infrastracture/model/circle_api_response/circle_api_response_model.dart';
+import 'package:m3_app/infrastructure/model/circle_api_response/circle_api_response_model.dart';
 import 'package:retrofit/retrofit.dart';
 
 part 'circle_api_service.g.dart';
