+++ lib/infrastructure/local/db/circle_database.dart
