+++ lib/provider/infrastructure/database_provider.dart
@@ -1,4 +1,4 @@
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:m3_app/infrastracture/local/db/circle_database.dart';
+import 'package:m3_app/infrastructure/local/db/circle_database.dart';
 
 final databaseProvider = Provider<CircleDatabase>((_) => CircleDatabase());
