+++ lib/provider/infrastructure/http_service_providers.dart
@@ -1,6 +1,6 @@
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
-import '../../infrastracture/http/circle_api/circle_api_service.dart';
+import '../../infrastructure/http/circle_api/circle_api_service.dart';
 import '../common/dio/dio_provider.dart';
 
 final circleApiServiceProvider = Provider<CircleApiService>((ref) {
